+++ packages/core/signal/src/core.ts
@@ -121,15 +121,10 @@ export function _addSignalListener<SignalName extends keyof AlwatrSignals>(
   if (signal.value !== undefined) {
     // null is a valid value for signal.
     if (options.receivePrevious === 'Immediate') {
-      logger.incident(
-          '_addSignalListener',
-          'call_signal_callback',
-          'run callback with previous signal value',
-          {
-            signalName: signal.name,
-            mode: 'Immediate',
-          },
-      );
+      logger.incident('_addSignalListener', 'call_signal_callback', 'run callback with previous signal value', {
+        signalName: signal.name,
+        mode: 'Immediate',
+      });
       try {
         listenerCallback(signal.value);
       }