+++ packages/core/storage-client/src/type.ts
@@ -30,7 +30,7 @@ export type AlwatrStorageClientConfig = {
    *
    * @default undefined Auto detect base on `NODE_ENV`
    */
-   debug?: boolean;
+  debug?: boolean;
 };
 
 type ServerResponseFailed = {