+++ packages/core/storage-client/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.22.1](https://github.com/AliMD/alwatr/compare/v0.22.0...v0.22.1) (2022-11-21)
+
+### Bug Fixes
+
+- **storage-client:** types issues ([c083bd0](https://github.com/AliMD/alwatr/commit/c083bd08cd2364bc64ebd1e60c668f37784d8502))
+
 # [0.22.0](https://github.com/AliMD/alwatr/compare/v0.21.0...v0.22.0) (2022-11-20)
 
 ### Bug Fixes